Update doxygen config.
Generate XML output. A thirdparty tool, such as breathe for sphinx, may
use the XML output to generate documentation in another format.

Add @algorithm tag. This tag may be used to start a section to document
the algorithm details (the "how"). On the other hand, the description
section should only describe the behavior of the function (the "what").