Commit work-in-progress of generic matrix operation refactoring.
The objective is to implement matrix operations on generic matrix-like
types.  However, it turns out difficult to make progress when starting
with generic code, and it proves hard to verify that the code actually
works for matrix-like types of interest.

Therefore, I choose to suspend this branch of work for the time being
and focus on a concrete implmentation, i.e. the euler::matrix class.